requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
ollama
Pillow